+++ modules/controller_build/versions.tf
@@ -4,5 +4,9 @@ terraform {
       source  = "hashicorp/azurerm"
       version = ">= 2.8.0"
     }
+    random = {
+      source  = "hashicorp/random"
+      version = ">= 3.0"
+    }
   }
 }